move device settings to main settings
